+++ src/components/VHeader/VPageMenu/VPageList.vue
@@ -1,16 +1,13 @@
 <template>
-  <VItemGroup
-    :direction="isMinScreenMd ? 'vertical' : 'columns'"
-    :class="{ 'mt-10': !isMinScreenMd }"
-    :bordered="false"
-  >
+  <VItemGroup :direction="layout" class="mt-10 md:mt-2" :bordered="false">
     <VItem
       v-for="(page, idx) in pages.all"
       :key="page.id"
-      :class="{ 'w-1/2': !isMinScreenMd }"
+      class="md:w-full w-1/2"
       :selected="page.id === pages.current"
       :is-first="idx === 0"
       v-bind="getLinkProps(page)"
+      @click="$emit('click')"
     >
       <div class="flex flex-row">
         <span class="pe-2">{{ $t(page.name) }}</span>
@@ -26,7 +23,6 @@
 </template>
 <script>
 import usePages from '~/composables/use-pages'
-import { isMinScreen } from '~/composables/use-media-query'
 
 import externalLinkIcon from '~/assets/icons/external-link.svg'
 
@@ -39,9 +35,15 @@ const externalLinkProps = { as: 'a', target: '_blank', rel: 'noopener' }
 export default {
   name: 'VPageMenuPopover',
   components: { VIcon, VItem, VItemGroup },
+  props: {
+    layout: {
+      type: String,
+      default: 'vertical',
+      validator: (val) => ['vertical', 'columns'].includes(val),
+    },
+  },
   setup() {
     const pages = usePages()
-    const isMinScreenMd = isMinScreen('md')
     const isLinkExternal = (item) => !item.link.startsWith('/')
     const getLinkProps = (item) => {
       return isLinkExternal(item)
@@ -53,7 +55,6 @@ export default {
       isLinkExternal,
       externalLinkIcon,
       pages,
-      isMinScreenMd,
     }
   },
 }
