+++ src/locales/scripts/wp-locales.json
@@ -385,7 +385,7 @@
     "facebookLocale": "da_DK",
     "wpLocale": "da_DK",
     "code": "da",
-    "translated": 98
+    "translated": 96
   },
   "de_AT": {
     "name": "German (Austria)",
@@ -467,7 +467,7 @@
     "facebookLocale": "el_GR",
     "wpLocale": "el",
     "code": "el",
-    "translated": 53
+    "translated": 51
   },
   "en_AU": {
     "name": "English (Australia)",
@@ -491,7 +491,7 @@
     "googleCode": "en",
     "wpLocale": "en_CA",
     "code": "en-ca",
-    "translated": 90
+    "translated": 89
   },
   "en_GB": {
     "name": "English (UK)",
@@ -528,7 +528,7 @@
     "googleCode": "en",
     "wpLocale": "en_ZA",
     "code": "en-za",
-    "translated": 95
+    "translated": 94
   },
   "eo": {
     "name": "Esperanto",
@@ -553,7 +553,7 @@
     "facebookLocale": "es_LA",
     "wpLocale": "es_AR",
     "code": "es-ar",
-    "translated": 94
+    "translated": 93
   },
   "es_CL": {
     "name": "Spanish (Chile)",
@@ -605,7 +605,7 @@
     "facebookLocale": "es_LA",
     "wpLocale": "es_DO",
     "code": "es-do",
-    "translated": 90
+    "translated": 89
   },
   "es_EC": {
     "name": "Spanish (Ecuador)",
@@ -670,7 +670,7 @@
     "facebookLocale": "es_MX",
     "wpLocale": "es_MX",
     "code": "es-mx",
-    "translated": 85
+    "translated": 100
   },
   "es_PE": {
     "name": "Spanish (Peru)",
@@ -848,7 +848,7 @@
     "facebookLocale": "fr_FR",
     "wpLocale": "fr_FR",
     "code": "fr",
-    "translated": 92
+    "translated": 90
   },
   "frp": {
     "name": "Arpitan",
@@ -941,7 +941,7 @@
     "facebookLocale": "gl_ES",
     "wpLocale": "gl_ES",
     "code": "gl",
-    "translated": 99
+    "translated": 98
   },
   "gu": {
     "name": "Gujarati",
@@ -1138,7 +1138,7 @@
     "facebookLocale": "it_IT",
     "wpLocale": "it_IT",
     "code": "it",
-    "translated": 47
+    "translated": 46
   },
   "ja": {
     "name": "Japanese",
@@ -1576,7 +1576,7 @@
     "googleCode": "nl",
     "wpLocale": "nl_BE",
     "code": "nl-be",
-    "translated": 37
+    "translated": 36
   },
   "nl_NL": {
     "name": "Dutch",
@@ -1589,7 +1589,7 @@
     "facebookLocale": "nl_NL",
     "wpLocale": "nl_NL",
     "code": "nl",
-    "translated": 100
+    "translated": 98
   },
   "nn_NO": {
     "name": "Norwegian (Nynorsk)",
@@ -1737,7 +1737,7 @@
     "facebookLocale": "pt_BR",
     "wpLocale": "pt_BR",
     "code": "pt-br",
-    "translated": 73
+    "translated": 72
   },
   "pt_PT": {
     "name": "Portuguese (Portugal)",
@@ -1773,7 +1773,7 @@
     "facebookLocale": "ro_RO",
     "wpLocale": "ro_RO",
     "code": "ro",
-    "translated": 100
+    "translated": 98
   },
   "roh": {
     "name": "Romansh",
@@ -1929,7 +1929,7 @@
     "facebookLocale": "sq_AL",
     "wpLocale": "sq",
     "code": "sq",
-    "translated": 99
+    "translated": 97
   },
   "sq_XK": {
     "name": "Shqip (Kosovo)",
@@ -2002,7 +2002,7 @@
     "facebookLocale": "sv_SE",
     "wpLocale": "sv_SE",
     "code": "sv",
-    "translated": 73
+    "translated": 75
   },
   "sw": {
     "name": "Swahili",
@@ -2335,7 +2335,7 @@
     "facebookLocale": "zh_CN",
     "wpLocale": "zh_CN",
     "code": "zh-cn",
-    "translated": 73
+    "translated": 72
   },
   "zh_HK": {
     "name": "Chinese (Hong Kong)",
@@ -2374,7 +2374,7 @@
     "facebookLocale": "zh_TW",
     "wpLocale": "zh_TW",
     "code": "zh-tw",
-    "translated": 29
+    "translated": 28
   },
   "zul": {
     "name": "Zulu",
